refactor(coordinator): use barrel imports in tests
