+++ packages/coordinator/test/tool-permission/denylist.test.ts
@@ -10,8 +10,13 @@ import {
 import { join } from "node:path";
 import { tmpdir, homedir } from "node:os";
 
-import { loadPolicy, checkPermission, type PolicyConfig } from "../../src/tool-permission/policy";
-import { logPermissionDecision, type AuditEntry } from "../../src/tool-permission/audit";
+import {
+  checkPermission,
+  loadPolicy,
+  logPermissionDecision,
+  type AuditEntry,
+  type PolicyConfig,
+} from "../../src/tool-permission";
 
 const TMP = join(tmpdir(), `openomni-perm-test-${process.pid}`);
 const POLICY_PATH = join(TMP, "tool-policy.json");
